REST: extract request handler
